djangogram/requirements.txt
Django==4.2.7
django-taggit==5.0.1
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0